--- a/kotlin/src/main/kotlin/net/jqwik/kotlin/api/CombinatorDsl.kt
+++ b/kotlin/src/main/kotlin/net/jqwik/kotlin/api/CombinatorDsl.kt
@@ -152,7 +152,7 @@ internal class ValueBindings {
 
             block()
         } finally {
-            current.remove()
+            current.set(null)
         }
     }
 }