## Overview

Currently, we have a lot of combinators in our kotlin code. They generally look like this:

```kt
fun current(): Arbitrary<String> {
    return combine(
        String.any(),
        String.any(),
        String.any(),
        String.any(),
        String.any(),
        String.any(),
        String.any(),
    ) { first, second, third, fourth, fifth, sixth, seventh ->
        "1: $first, 2: $second, 3: $third, 4: $fourth, 5: $fifth, 6: $sixth, 7: $seventh"
    }
}
```

This has two major disadvantages in my opinion:
- It is hard to read and modify, because you don't directly know which arbitrary belongs to which value
- This only works with up to eight Arbitraries, because of the way the combinator functions have to be written

The syntax that is added by this PR allows to write a combinator like this:

```kt
fun proposed(): Arbitrary<String> {
    return combine {
        val first by String.any()
        val second by String.any()
        val third by String.any()
        val fourth by String.any()
        val fifth by String.any()
        val sixth by String.any()
        val seventh by String.any()

        createAs {
            "1: $first, 2: $second, 3: $third, 4: $fourth, 5: $fifth, 6: $sixth, 7: $seventh"
        }
    }
}
```

### Details

I just wanted to create a PR in case you like this DSL, because I think it would be a cool addition. The actual implementation is somewhat hacky, so I can imagine that it might not be something you want. Otherwise I'll just keep it in our fork we use internally 😉

---

I hereby agree to the terms of the [jqwik Contributor Agreement](https://github.com/jqwik-team/jqwik/blob/master/CONTRIBUTING.md#jqwik-contributor-agreement).