@fxshlein 2023-08-03
> This might be relevant as well: [#329 (comment)](https://github.com/jqwik-team/jqwik/issues/329#issuecomment-1273322613)

Good point, we don't really use filtering that much since we try to just generate the arguments, but I can imagine something like this would work within the DSL:
```kt
fun proposed(): Arbitrary<String> {
    return combine {
        val first by String.any()
        val second by String.any()

       filter { 'a' !in first }
       filter { 'b' !in second }
       filter { first != second }
  
        createAs {
            "1: $first, 2: $second"
        }
    }
}
```

Or perhaps, instead of `filter` something more "expressive" like `where` or `require`?