@vlsi 2023-08-03
>Or perhaps, instead of filter something more "expressive" like where or require?

Yeah. I don't ask for including "filter / where" into the current PR, however, it would be nice if the concept could be enhanced to cover more cases in the future.

There's `flatCombine` as well which works as follows (`createAs` returns `Arbitrary`)

```kotlin
fun <T1, T2, R> flatCombine(
    a1: Arbitrary<T1>,
    a2: Arbitrary<T2>,
    flatCombinator: (v1: T1, v2: T2) -> Arbitrary<R>
): Arbitrary<R>
```