@fxshlein 2023-08-03
@vlsi I pushed a bunch of changes/suggestions:
- Thread local is now created once per `combine` call
- Added `filter` and `flatCombine` (since it seemed pretty simple to add)
- Added an exception that is thrown when calling `createAs` multiple times