@fxshlein 2023-08-03
I also removed the `CombinatorCreateScope`, I realized it wasn't actually needed, it was from an approach I experimented with before