@jlink 2023-08-04
It's a clever solution. Since I'm not doing a lot of Kotlin myself I don't feel ready to judge if the slight improvement of readability (my subjective feeling) is worth introducing yet another combinator DSL; there are already three: `Combinators.combine(..)`, `Builders` and Kotlin extension functions `combine(..)`.

It comes with a performance hit:

```kotlin
@Example
fun `measure performance`() {
    val combinedNew = combine {
        val v1 by Int.any(1..1000)
        val v2 by Int.any(1..1000)

        createAs {
            v1!! + v2!!
        }
    }

    val repetitions: Long = 10_000_000

    val timeNew = measureTime {
        combinedNew.sampleStream().limit(repetitions).forEach { it }
    }
    println("new approach: $timeNew")

    val combinedOld = combine(Int.any(1..1000), Int.any(1..1000)) { v1, v2 -> v1 + v2 }
    val timeOld = measureTime {
        combinedOld.sampleStream().limit(repetitions).forEach { it }
    }
    println("old approach: $timeOld")
}
```

```
new approach: 3.008442949s
old approach: 1.828261633s
```

All code is a liability ;-/