@vlsi 2023-08-04
Regarding `ThreadLocal` removal. It looks like delegated properties can't access extra context at the use-site, and I guess the issue is known as https://youtrack.jetbrains.com/issue/KT-10468/Context-receivers-multiple-receivers-on-extension-functions-properties and https://youtrack.jetbrains.com/issue/KT-14504/Support-delegated-properties-with-second-receiver-so-that-they-are-available-only-in-a-certain-context-i.e.-STM .

So the workaround could be adding an extra `.value` which will exist in `BindScope` only:

```kotlin
fun proposed(): Arbitrary<String> {
    return combine {
        val first by String.any()
        val second by String.any()
        val third by String.any()
        val fourth by String.any()
        val fifth by String.any()
        val sixth by String.any()
        val seventh by String.any()

        createAs { // this: BindScope
             // Imagine class BindScope { val PropType.value:  ...
            "1: ${first.value}, 2: ${second.value}, 3: ${third.value}, 4..."
        }
    }
}
```

The explicit `.value` would effectively capture `BindScope`, so `ThreadLocal` would not be needed for passing the values.

An alternative option could be adding a function in `BindScope` context:
For instance:

```kotlin
        createAs { // this: BindScope
             // Imagine class BindScope { operator fun PropType.invoke() =  ...
            "1: ${first()}, 2: ${second()}, 3: ${third()}, 4..."
        }
```

WDYT?

The approach with extra function would likely be Java-friendly as well.

Java syntax could be

```java
var combined = combine((combineScope) -> {
    var first = combineScope.registerVariable(Arbitraries.strings());
    var second = combineScope.registerVariable(Arbitraries.ints());
    combineScope.combineAs((bindScope) ->
        first.getValue(bindScope) + second.getValue(bindScope)
        // or bindScope.get(first) + bindScope.get(second) if it looks better
    });
});
```

In that case, `ThreadLocals` are not needed, and the same API could more-or-less work for both Java and Kotlin.

Of course, as Kotlin implements context receivers, it might provide a nicer API without explicit `.value`.

WDYT?