@jlink 2023-08-04
> An alternative option could be adding a function in BindScope context:
> For instance:
```
createAs { // this: BindScope
             // Imagine class BindScope { operator fun PropType.invoke() =  ...
            "1: ${first()}, 2: ${second()}, 3: ${third()}, 4..."
        }
```

Couldn't the property's getter function bind to the scope?
I'm still hoping there's a way to stick with property access without using a thread local.