@jlink 2023-08-04
> The approach with extra function would likely be Java-friendly as well.

That has no importance to me. Feels completely unnatural in Java.