@fxshlein 2023-08-04
> Couldn't the property's getter function bind to the scope?

I've been experimenting with this, but the issue is that kotlin takes whatever `getValue` function it finds at the place where the property is delegated (so essentially wherever the `by` keyword is) and then runs with that, without a way to overwrite that.