@fxshlein 2023-08-04
Essentially, kotlin needs a `.getValue()` function that works inside `combine`:
```kotlin
        val combined = combine {
            // Kotlin determines which 'ArbitraryProperty<T>.getValue()'
            // function it will use here, so it has to work here:
            val value by just(1)

            combineAs {
                // So even if another ' getValue' is available in this scope, it is ignored.
                value
            }
        }
```

For that reason, you cannot bind it to the scope in `combineAs`, because it will always have to bind to the scope in `combine`.

If there is no `.getValue()` at all available in `combine`, kotlin will complain - so only making it available in `combineAs` is also not possible.