@jlink 2023-08-04
> If there is no `.getValue()` at all available in `combine`, kotlin will complain - so only making it available in `combineAs` is also not possible.

I wonder if one of Kotlin's scope functions could make a difference, like:
```kotlin
val combined = combine {
            // Kotlin determines which 'ArbitraryProperty<T>.getValue()'
            // function it will use here, so it has to work here:
            val value by just(1)

            combineAs {
                bindings.let {
                     value // would it use value in bindings?
                }
            }
        }
```