@fxshlein 2023-08-04
> > If there is no `.getValue()` at all available in `combine`, kotlin will complain - so only making it available in `combineAs` is also not possible.
> 
> I wonder if one of Kotlin's scope functions could make a difference, like:

Sadly not, because they also just use scopes and receivers in the same way we can use them.