@fxshlein 2023-08-04
I think there is no way around having either the `ThreadLocal`, or not having the property access syntax.

I profiled a bit btw, and a lot of the performance drop actually comes from the `ThreadLocal.remove()`?! I wonder if it would be fine and speed stuff up to just replace that with `set(null)`