@fxshlein 2023-08-04
I've done some more benchmarking. The test you sent above on my machine:
```
new approach: 2.482132500s
old approach: 1.483676100s
```

However, replacing `ThreadLocal.remove()` with `.set(null)`, I get this:
```
new approach: 1.942857400s
old approach: 1.645651901s
```

That's quite the big difference, but I think I figured out why: Since ThreadLocal.remove()` will completely remove the value for that Thread from the map of thread local values, the next call to `.set(values)` will have to make a completely new entry in the map. If I only call `.set(null)` however, I still remove the reference to the values so they can be garbage collected, but the next `.set(values)` call will only have to update the existing entry that is still set to null, instead of having to create a completely new entry. So it makes it more performant on both ends.